backend/requirements.txt
fastapi
uvicorn
numpy
pandas
faiss-cpu
sqlalchemy
pydantic
python-dotenv
python-jose
passlib